chore: move test to options_test.go
