+++ options_test.go
@@ -12,7 +12,62 @@ import (
 
 	"github.com/testcontainers/testcontainers-go"
 	"github.com/testcontainers/testcontainers-go/exec"
+	"github.com/testcontainers/testcontainers-go/wait"
 )
+
+func TestOverrideContainerRequest(t *testing.T) {
+	req := testcontainers.GenericContainerRequest{
+		ContainerRequest: testcontainers.ContainerRequest{
+			Env: map[string]string{
+				"BAR": "BAR",
+			},
+			Image:        "foo",
+			ExposedPorts: []string{"12345/tcp"},
+			WaitingFor: wait.ForNop(
+				func(ctx context.Context, target wait.StrategyTarget) error {
+					return nil
+				},
+			),
+			Networks: []string{"foo", "bar", "baaz"},
+			NetworkAliases: map[string][]string{
+				"foo": {"foo0", "foo1", "foo2", "foo3"},
+			},
+		},
+	}
+
+	toBeMergedRequest := testcontainers.GenericContainerRequest{
+		ContainerRequest: testcontainers.ContainerRequest{
+			Env: map[string]string{
+				"FOO": "FOO",
+			},
+			Image:        "bar",
+			ExposedPorts: []string{"67890/tcp"},
+			Networks:     []string{"foo1", "bar1"},
+			NetworkAliases: map[string][]string{
+				"foo1": {"bar"},
+			},
+			WaitingFor: wait.ForLog("foo"),
+		},
+	}
+
+	// the toBeMergedRequest should be merged into the req
+	testcontainers.CustomizeRequest(toBeMergedRequest)(&req)
+
+	// toBeMergedRequest should not be changed
+	assert.Equal(t, "", toBeMergedRequest.Env["BAR"])
+	assert.Equal(t, 1, len(toBeMergedRequest.ExposedPorts))
+	assert.Equal(t, "67890/tcp", toBeMergedRequest.ExposedPorts[0])
+
+	// req should be merged with toBeMergedRequest
+	assert.Equal(t, "FOO", req.Env["FOO"])
+	assert.Equal(t, "BAR", req.Env["BAR"])
+	assert.Equal(t, "bar", req.Image)
+	assert.Equal(t, []string{"12345/tcp", "67890/tcp"}, req.ExposedPorts)
+	assert.Equal(t, []string{"foo", "bar", "baaz", "foo1", "bar1"}, req.Networks)
+	assert.Equal(t, []string{"foo0", "foo1", "foo2", "foo3"}, req.NetworkAliases["foo"])
+	assert.Equal(t, []string{"bar"}, req.NetworkAliases["foo1"])
+	assert.Equal(t, wait.ForLog("foo"), req.WaitingFor)
+}
 
 func TestWithNetwork(t *testing.T) {
 	req := testcontainers.GenericContainerRequest{
